requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
automium==0.2.6
automium_web==0.1.1
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
distlib==0.3.9
Django==5.2.3
django-ratelimit==4.1.0
django-ratelimit2==0.4.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
filelock==3.18.0
Flask==3.1.1
humansignal-drf-yasg==1.21.10.post1
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
legacy==0.1.7
MarkupSafe==3.0.2
openapi==2.0.0
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
PyJWT==2.9.0
pyTelegramBotAPI==4.27.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
six==1.17.0
sqlparse==0.5.3
telebot==0.0.5
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.13
Werkzeug==3.1.3
